requirements.txt
APScheduler==3.10.1
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pycodestyle==2.10.0
pyTelegramBotAPI==4.10.0
python-decouple==3.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
soupsieve==2.4
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
